Changed DB schema + bug fix
DB schema now makes use of the default data column, which is required anyway. Fixed a bug where an empty weather object would be returned and stored if the API request failed. If available, the outdated value will now be returned if a a new one couldn’t be retrieved. Otherwise it will be null so it’s easy to check.